The contrib/ydb/public/sdk/cpp/client was removed from implicit dependencies through nested #include
